+++ src/renderer/components/common/Button/Button.css
@@ -1,11 +1,13 @@
 .button {
+  align-items: center;
   appearance: none;
   background-color: var(--ba-light);
   border: var(--border-1) solid var(--ba-border-active);
   border-radius: var(--margin-025);
+  display: flex;
   color: var(--ba-dark);
   flex: 1;
-  flex-direction: column;
+  /* flex-direction: column; */
   font-family: inherit;
   font-size: 0.95rem;
   line-height: inherit;
@@ -27,10 +29,20 @@
   outline: none;
 }
 
+.button--style-default {
+  /* display: flex; */
+}
+
 .button__icon-container {
-  width: 100%;
+  display: flex;
+  align-items: center;
+  margin-right: var(--margin-050);
+  /* width: 100%; */
+  /* display: inline; */
 }
 
 .button__label {
-  width: 100%;
+  /* width: 100%; */
+  /* display: inline; */
+  white-space: nowrap;
 }
