+++ src/renderer/components/common/ModalFooter/ModalFooter.css
@@ -1,6 +1,8 @@
 .modal-footer {
   border-top: var(--border-1) solid var(--ba-border);
-  height: var(--height-400);
+  display: flex;
+  /* height: var(--height-500); */
+  justify-content: space-between;
   margin: 0;
   padding: var(--margin-100);
 }
